Update Autoformer ETTm1.sh script with --freq 't'
While the Dataset_ETT_minute(Dataset) has as default 'h' in the freq argument in the class constructor,when  I execute the ETTm1 script the freq argument is overridden with the default which is --freq 'h' and if I am not wrong it is also passed to the __init__ of the class Dataset_ETT_minute(Dataset). Am I wrong?